Release 4.8.0 (#651)
- [API] Deprecate all gQUIC versions.
- [FIX] http1x: check size of Cookie: header.
- [FIX] Add checks to gQUIC client session resume code.
- [FIX] 32-bit parser bounds checks.
- [FIX] Use ls-qpack v2.6.3 to pick up a bunch of fixes.
- [DOCS] FAQ: example programs do work with google.com etc.

+++ src/liblsquic/lsquic_trans_params.c
@@ -535,7 +535,7 @@ lsquic_tp_decode (const unsigned char *const buf, size_t bufsz,
         if (s < 0)
             return -1;
         p += s;
-        if ((ptrdiff_t) len > end - p)
+        if (len > (uint64_t) (end - p))
             return -1;
         tpi = tpi_val_2_enum(param_id);
         if (tpi <= LAST_TPI)
@@ -1125,7 +1125,7 @@ lsquic_tp_decode_27 (const unsigned char *const buf, size_t bufsz,
         if (s < 0)
             return -1;
         p += s;
-        if ((ptrdiff_t) len > end - p)
+        if (len > (uint64_t) (end - p))
             return -1;
         tpi = tpi_val_2_enum(param_id);
         if (tpi <= LAST_TPI)
